geotab-mcp-docker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
mygeotab==0.8.8
python-dotenv==1.0.0
requests==2.31.0
